refactor(backend): split media pipeline actions
